requirements.txt
flask
flask-cors
PyJWT
bcrypt
psycopg2-binary
python-dotenv
